move common structures from fisheye.cpp to fisheye.hpp
