+++ modules/calib3d/src/fisheye.cpp
@@ -44,19 +44,6 @@
 #include "fisheye.hpp"
 #include <limits>
 
-namespace cv { namespace
-{
-    struct JacobianRow
-    {
-        Vec2d df, dc;
-        Vec4d dk;
-        Vec3d dom, dT;
-        double dalpha;
-    };
-
-    void subMatrix(const Mat& src, Mat& dst, const std::vector<uchar>& cols, const std::vector<uchar>& rows);
-}}
-
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /// cv::fisheye::projectPoints
